+++ src/platform_impl/ios/ffi.rs
@@ -2,6 +2,7 @@
 
 use std::{convert::TryInto, ffi::CString, ops::BitOr, os::raw::*};
 
+use objc::foundation::{NSInteger, NSUInteger};
 use objc::{runtime::Object, Encode, Encoding};
 
 use crate::{
@@ -17,9 +18,6 @@ pub type CGFloat = f32;
 #[cfg(target_pointer_width = "64")]
 pub type CGFloat = f64;
 
-pub type NSInteger = isize;
-pub type NSUInteger = usize;
-
 #[repr(C)]
 #[derive(Clone, Debug)]
 pub struct NSOperatingSystemVersion {
@@ -28,11 +26,26 @@ pub struct NSOperatingSystemVersion {
     pub patch: NSInteger,
 }
 
+unsafe impl Encode for NSOperatingSystemVersion {
+    const ENCODING: Encoding = Encoding::Struct(
+        "NSOperatingSystemVersion",
+        &[
+            NSInteger::ENCODING,
+            NSInteger::ENCODING,
+            NSInteger::ENCODING,
+        ],
+    );
+}
+
 #[repr(C)]
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct CGPoint {
     pub x: CGFloat,
     pub y: CGFloat,
+}
+
+unsafe impl Encode for CGPoint {
+    const ENCODING: Encoding = Encoding::Struct("CGPoint", &[CGFloat::ENCODING, CGFloat::ENCODING]);
 }
 
 #[repr(C)]
@@ -51,6 +64,10 @@ impl CGSize {
     }
 }
 
+unsafe impl Encode for CGSize {
+    const ENCODING: Encoding = Encoding::Struct("CGSize", &[CGFloat::ENCODING, CGFloat::ENCODING]);
+}
+
 #[repr(C)]
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct CGRect {
@@ -65,18 +82,9 @@ impl CGRect {
 }
 
 unsafe impl Encode for CGRect {
-    fn encode() -> Encoding {
-        unsafe {
-            if cfg!(target_pointer_width = "32") {
-                Encoding::from_str("{CGRect={CGPoint=ff}{CGSize=ff}}")
-            } else if cfg!(target_pointer_width = "64") {
-                Encoding::from_str("{CGRect={CGPoint=dd}{CGSize=dd}}")
-            } else {
-                unimplemented!()
-            }
-        }
-    }
+    const ENCODING: Encoding = Encoding::Struct("CGRect", &[CGPoint::ENCODING, CGSize::ENCODING]);
 }
+
 #[derive(Debug)]
 #[allow(dead_code)]
 #[repr(isize)]
@@ -88,6 +96,10 @@ pub enum UITouchPhase {
     Cancelled,
 }
 
+unsafe impl Encode for UITouchPhase {
+    const ENCODING: Encoding = NSInteger::ENCODING;
+}
+
 #[derive(Debug, PartialEq, Eq)]
 #[allow(dead_code)]
 #[repr(isize)]
@@ -95,6 +107,10 @@ pub enum UIForceTouchCapability {
     Unknown = 0,
     Unavailable,
     Available,
+}
+
+unsafe impl Encode for UIForceTouchCapability {
+    const ENCODING: Encoding = NSInteger::ENCODING;
 }
 
 #[derive(Debug, PartialEq, Eq)]
@@ -106,6 +122,10 @@ pub enum UITouchType {
     Pencil,
 }
 
+unsafe impl Encode for UITouchType {
+    const ENCODING: Encoding = NSInteger::ENCODING;
+}
+
 #[repr(C)]
 #[derive(Debug, Clone)]
 pub struct UIEdgeInsets {
@@ -115,14 +135,24 @@ pub struct UIEdgeInsets {
     pub right: CGFloat,
 }
 
+unsafe impl Encode for UIEdgeInsets {
+    const ENCODING: Encoding = Encoding::Struct(
+        "UIEdgeInsets",
+        &[
+            CGFloat::ENCODING,
+            CGFloat::ENCODING,
+            CGFloat::ENCODING,
+            CGFloat::ENCODING,
+        ],
+    );
+}
+
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct UIUserInterfaceIdiom(NSInteger);
 
 unsafe impl Encode for UIUserInterfaceIdiom {
-    fn encode() -> Encoding {
-        NSInteger::encode()
-    }
+    const ENCODING: Encoding = NSInteger::ENCODING;
 }
 
 impl UIUserInterfaceIdiom {
@@ -162,9 +192,7 @@ impl From<UIUserInterfaceIdiom> for Idiom {
 pub struct UIInterfaceOrientationMask(NSUInteger);
 
 unsafe impl Encode for UIInterfaceOrientationMask {
-    fn encode() -> Encoding {
-        NSUInteger::encode()
-    }
+    const ENCODING: Encoding = NSUInteger::ENCODING;
 }
 
 impl UIInterfaceOrientationMask {
@@ -213,9 +241,7 @@ impl UIInterfaceOrientationMask {
 pub struct UIRectEdge(NSUInteger);
 
 unsafe impl Encode for UIRectEdge {
-    fn encode() -> Encoding {
-        NSUInteger::encode()
-    }
+    const ENCODING: Encoding = NSUInteger::ENCODING;
 }
 
 impl From<ScreenEdge> for UIRectEdge {
@@ -241,9 +267,7 @@ impl From<UIRectEdge> for ScreenEdge {
 pub struct UIScreenOverscanCompensation(NSInteger);
 
 unsafe impl Encode for UIScreenOverscanCompensation {
-    fn encode() -> Encoding {
-        NSInteger::encode()
-    }
+    const ENCODING: Encoding = NSInteger::ENCODING;
 }
 
 #[allow(dead_code)]