+++ src/platform_impl/ios/mod.rs
@@ -63,8 +63,7 @@
 // window size/position.
 macro_rules! assert_main_thread {
     ($($t:tt)*) => {
-        let is_main_thread: ::objc::runtime::BOOL = msg_send!(class!(NSThread), isMainThread);
-        if is_main_thread == ::objc::runtime::NO {
+        if !::objc::foundation::is_main_thread() {
             panic!($($t)*);
         }
     };