+++ src/platform_impl/macos/window.rs
@@ -5,7 +5,7 @@ use std::{
     os::raw::c_void,
     sync::{
         atomic::{AtomicBool, Ordering},
-        Arc, Mutex, MutexGuard, Weak,
+        Arc, Mutex, MutexGuard,
     },
 };
 
@@ -26,8 +26,7 @@ use crate::{
         ffi,
         monitor::{self, MonitorHandle, VideoMode},
         util::{self, IdRef},
-        view::CursorState,
-        view::{self, new_view},
+        view::{self, new_view, ViewState},
         window_delegate::new_delegate,
         OsError,
     },
@@ -42,15 +41,15 @@ use cocoa::{
         NSRequestUserAttentionType, NSScreen, NSView, NSWindow, NSWindowButton, NSWindowStyleMask,
     },
     base::{id, nil},
-    foundation::{NSArray, NSDictionary, NSPoint, NSRect, NSSize, NSString, NSUInteger},
+    foundation::{NSDictionary, NSPoint, NSRect, NSSize, NSString, NSArray},
 };
 use core_graphics::display::{CGDisplay, CGDisplayMode};
-use objc::{
-    declare::ClassDecl,
-    rc::autoreleasepool,
-    runtime::{Class, Object, Sel, BOOL, NO, YES},
-};
-use once_cell::sync::Lazy;
+use objc2::foundation::{is_main_thread, NSObject, NSUInteger};
+use objc2::rc::autoreleasepool;
+use objc2::runtime::{Bool, Object};
+use objc2::{declare_class, ClassType};
+
+use super::appkit::{NSCursor, NSResponder, NSWindow as NSWindowClass};
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct WindowId(pub usize);
@@ -112,16 +111,16 @@ impl Default for PlatformSpecificWindowBuilderAttributes {
 unsafe fn create_view(
     ns_window: id,
     pl_attribs: &PlatformSpecificWindowBuilderAttributes,
-) -> Option<(IdRef, Weak<Mutex<CursorState>>)> {
-    let (ns_view, cursor_state) = new_view(ns_window);
+) -> Option<IdRef> {
+    let ns_view = new_view(ns_window);
     ns_view.non_nil().map(|ns_view| {
         // The default value of `setWantsBestResolutionOpenGLSurface:` was `false` until
         // macos 10.14 and `true` after 10.15, we should set it to `YES` or `NO` to avoid
         // always the default system value in favour of the user's code
         if !pl_attribs.disallow_hidpi {
-            ns_view.setWantsBestResolutionOpenGLSurface_(YES);
+            ns_view.setWantsBestResolutionOpenGLSurface_(Bool::YES.as_raw());
         } else {
-            ns_view.setWantsBestResolutionOpenGLSurface_(NO);
+            ns_view.setWantsBestResolutionOpenGLSurface_(Bool::NO.as_raw());
         }
 
         // On Mojave, views automatically become layer-backed shortly after being added to
@@ -130,10 +129,10 @@ unsafe fn create_view(
         // explicitly make the view layer-backed up front so that AppKit doesn't do it
         // itself and break the association with its context.
         if f64::floor(appkit::NSAppKitVersionNumber) > appkit::NSAppKitVersionNumber10_12 {
-            ns_view.setWantsLayer(YES);
+            ns_view.setWantsLayer(Bool::YES.as_raw());
         }
 
-        (ns_view, cursor_state)
+        ns_view
     })
 }
 
@@ -141,7 +140,7 @@ fn create_window(
     attrs: &WindowAttributes,
     pl_attrs: &PlatformSpecificWindowBuilderAttributes,
 ) -> Option<IdRef> {
-    autoreleasepool(|| unsafe {
+    autoreleasepool(|_| unsafe {
         let screen = match attrs.fullscreen {
             Some(Fullscreen::Borderless(Some(RootMonitorHandle { inner: ref monitor })))
             | Some(Fullscreen::Exclusive(RootVideoMode {
@@ -203,22 +202,22 @@ fn create_window(
             masks |= NSWindowStyleMask::NSFullSizeContentViewWindowMask;
         }
 
-        let ns_window: id = msg_send![WINDOW_CLASS.0, alloc];
+        let ns_window: id = msg_send![WinitWindow::class(), alloc];
         let ns_window = IdRef::new(ns_window.initWithContentRect_styleMask_backing_defer_(
             frame,
             masks,
             appkit::NSBackingStoreBuffered,
-            NO,
+            Bool::NO.as_raw(),
         ));
 
         ns_window.non_nil().map(|ns_window| {
             let title = util::ns_string_id_ref(&attrs.title);
-            ns_window.setReleasedWhenClosed_(NO);
+            ns_window.setReleasedWhenClosed_(Bool::NO.as_raw());
             ns_window.setTitle_(*title);
-            ns_window.setAcceptsMouseMovedEvents_(YES);
+            ns_window.setAcceptsMouseMovedEvents_(Bool::YES.as_raw());
 
             if pl_attrs.titlebar_transparent {
-                ns_window.setTitlebarAppearsTransparent_(YES);
+                ns_window.setTitlebarAppearsTransparent_(Bool::YES.as_raw());
             }
             if pl_attrs.title_hidden {
                 ns_window.setTitleVisibility_(appkit::NSWindowTitleVisibility::NSWindowTitleHidden);
@@ -231,18 +230,15 @@ fn create_window(
                     NSWindowButton::NSWindowZoomButton,
                 ] {
                     let button = ns_window.standardWindowButton_(*titlebar_button);
-                    let _: () = msg_send![button, setHidden: YES];
+                    let _: () = msg_send![button, setHidden: true];
                 }
             }
             if pl_attrs.movable_by_window_background {
-                ns_window.setMovableByWindowBackground_(YES);
+                ns_window.setMovableByWindowBackground_(Bool::YES.as_raw());
             }
 
             if attrs.always_on_top {
-                let _: () = msg_send![
-                    *ns_window,
-                    setLevel: ffi::NSWindowLevel::NSFloatingWindowLevel
-                ];
+                let _: () = msg_send![*ns_window, setLevel: ffi::kCGFloatingWindowLevelKey];
             }
 
             if let Some(increments) = pl_attrs.resize_increments {
@@ -254,7 +250,7 @@ fn create_window(
             }
 
             if !pl_attrs.has_shadow {
-                ns_window.setHasShadow_(NO);
+                ns_window.setHasShadow_(Bool::NO.as_raw());
             }
             if attrs.position.is_none() {
                 ns_window.center();
@@ -264,34 +260,28 @@ fn create_window(
     })
 }
 
-struct WindowClass(*const Class);
-unsafe impl Send for WindowClass {}
-unsafe impl Sync for WindowClass {}
+declare_class!(
+    struct WinitWindow {}
 
-static WINDOW_CLASS: Lazy<WindowClass> = Lazy::new(|| unsafe {
-    let window_superclass = class!(NSWindow);
-    let mut decl = ClassDecl::new("WinitWindow", window_superclass).unwrap();
-
-    pub extern "C" fn can_become_main_window(_: &Object, _: Sel) -> BOOL {
-        trace_scope!("canBecomeMainWindow");
-        YES
+    unsafe impl ClassType for WinitWindow {
+        #[inherits(NSResponder, NSObject)]
+        type Super = NSWindowClass;
     }
 
-    pub extern "C" fn can_become_key_window(_: &Object, _: Sel) -> BOOL {
-        trace_scope!("canBecomeKeyWindow");
-        YES
-    }
+    unsafe impl WinitWindow {
+        #[sel(canBecomeMainWindow)]
+        fn can_become_main_window(&self) -> bool {
+            trace_scope!("canBecomeMainWindow");
+            true
+        }
 
-    decl.add_method(
-        sel!(canBecomeMainWindow),
-        can_become_main_window as extern "C" fn(&Object, Sel) -> BOOL,
-    );
-    decl.add_method(
-        sel!(canBecomeKeyWindow),
-        can_become_key_window as extern "C" fn(&Object, Sel) -> BOOL,
-    );
-    WindowClass(decl.register())
-});
+        #[sel(canBecomeKeyWindow)]
+        fn can_become_key_window(&self) -> bool {
+            trace_scope!("canBecomeKeyWindow");
+            true
+        }
+    }
+);
 
 #[derive(Default)]
 pub struct SharedState {
@@ -385,7 +375,6 @@ pub struct UnownedWindow {
     pub ns_view: IdRef,   // never changes
     pub(super) shared_state: Arc<Mutex<SharedState>>,
     decorations: AtomicBool,
-    cursor_state: Weak<Mutex<CursorState>>,
     pub inner_rect: Option<PhysicalSize<u32>>,
 }
 
@@ -397,18 +386,15 @@ impl UnownedWindow {
         mut win_attribs: WindowAttributes,
         pl_attribs: PlatformSpecificWindowBuilderAttributes,
     ) -> Result<(Arc<Self>, IdRef), RootOsError> {
-        unsafe {
-            let is_main_thread: BOOL = msg_send!(class!(NSThread), isMainThread);
-            if is_main_thread == NO {
-                panic!("Windows can only be created on the main thread on macOS");
-            }
+        if !is_main_thread() {
+            panic!("Windows can only be created on the main thread on macOS");
         }
         trace!("Creating new window");
 
         let ns_window = create_window(&win_attribs, &pl_attribs)
             .ok_or_else(|| os_error!(OsError::CreationError("Couldn't create `NSWindow`")))?;
 
-        let (ns_view, cursor_state) = unsafe { create_view(*ns_window, &pl_attribs) }
+        let ns_view = unsafe { create_view(*ns_window, &pl_attribs) }
             .ok_or_else(|| os_error!(OsError::CreationError("Couldn't create `NSView`")))?;
 
         // Configure the new view as the "key view" for the window
@@ -421,7 +407,7 @@ impl UnownedWindow {
 
         unsafe {
             if win_attribs.transparent {
-                ns_window.setOpaque_(NO);
+                ns_window.setOpaque_(Bool::NO.as_raw());
                 ns_window.setBackgroundColor_(NSColor::clearColor(nil));
             }
 
@@ -461,7 +447,6 @@ impl UnownedWindow {
             ns_window,
             shared_state: Arc::new(Mutex::new(win_attribs.into())),
             decorations: AtomicBool::new(decorations),
-            cursor_state,
             inner_rect,
         });
 
@@ -532,8 +517,8 @@ impl UnownedWindow {
 
     #[inline]
     pub fn is_visible(&self) -> Option<bool> {
-        let is_visible: BOOL = unsafe { msg_send![*self.ns_window, isVisible] };
-        Some(is_visible == YES)
+        let is_visible = unsafe { msg_send![*self.ns_window, isVisible] };
+        Some(is_visible)
     }
 
     pub fn request_redraw(&self) {
@@ -638,18 +623,22 @@ impl UnownedWindow {
 
     #[inline]
     pub fn is_resizable(&self) -> bool {
-        let is_resizable: BOOL = unsafe { msg_send![*self.ns_window, isResizable] };
-        is_resizable == YES
+        unsafe { msg_send![*self.ns_window, isResizable] }
     }
 
-    pub fn set_cursor_icon(&self, cursor: CursorIcon) {
-        let cursor = util::Cursor::from(cursor);
-        if let Some(cursor_access) = self.cursor_state.upgrade() {
-            cursor_access.lock().unwrap().cursor = cursor;
-        }
+    pub fn set_cursor_icon(&self, icon: CursorIcon) {
+        let view_state: &ViewState = unsafe {
+            let ns_view: &Object = (*self.ns_view).as_ref().expect("failed to deref");
+            let state_ptr: *const c_void = *ns_view.ivar("winitState");
+            &*(state_ptr as *const ViewState)
+        };
+        let mut cursor_state = view_state.cursor_state.lock().unwrap();
+        cursor_state.cursor = NSCursor::from_icon(icon);
+        drop(cursor_state);
         unsafe {
-            let _: () = msg_send![*self.ns_window,
-                invalidateCursorRectsForView:*self.ns_view
+            let _: () = msg_send![
+                *self.ns_window,
+                invalidateCursorRectsForView: *self.ns_view,
             ];
         }
     }
@@ -671,16 +660,19 @@ impl UnownedWindow {
 
     #[inline]
     pub fn set_cursor_visible(&self, visible: bool) {
-        if let Some(cursor_access) = self.cursor_state.upgrade() {
-            let mut cursor_state = cursor_access.lock().unwrap();
-            if visible != cursor_state.visible {
-                cursor_state.visible = visible;
-                drop(cursor_state);
-                unsafe {
-                    let _: () = msg_send![*self.ns_window,
-                        invalidateCursorRectsForView:*self.ns_view
-                    ];
-                }
+        let view_state: &ViewState = unsafe {
+            let ns_view: &Object = (*self.ns_view).as_ref().expect("failed to deref");
+            let state_ptr: *const c_void = *ns_view.ivar("winitState");
+            &*(state_ptr as *const ViewState)
+        };
+        let mut cursor_state = view_state.cursor_state.lock().unwrap();
+        if visible != cursor_state.visible {
+            cursor_state.visible = visible;
+            drop(cursor_state);
+            unsafe {
+                let _: () = msg_send![*self.ns_window,
+                    invalidateCursorRectsForView:*self.ns_view
+                ];
             }
         }
     }
@@ -739,14 +731,14 @@ impl UnownedWindow {
             self.set_style_mask_sync(required);
         }
 
-        let is_zoomed: BOOL = unsafe { msg_send![*self.ns_window, isZoomed] };
+        let is_zoomed = unsafe { msg_send![*self.ns_window, isZoomed] };
 
         // Roll back temp styles
         if needs_temp_mask {
             self.set_style_mask_async(curr_mask);
         }
 
-        is_zoomed != NO
+        is_zoomed
     }
 
     fn saved_style(&self, shared_state: &mut SharedState) -> NSWindowStyleMask {
@@ -780,8 +772,7 @@ impl UnownedWindow {
 
     #[inline]
     pub fn set_minimized(&self, minimized: bool) {
-        let is_minimized: BOOL = unsafe { msg_send![*self.ns_window, isMiniaturized] };
-        let is_minimized: bool = is_minimized == YES;
+        let is_minimized: bool = unsafe { msg_send![*self.ns_window, isMiniaturized] };
         if is_minimized == minimized {
             return;
         }
@@ -1011,10 +1002,7 @@ impl UnownedWindow {
 
                 // Restore the normal window level following the Borderless fullscreen
                 // `CGShieldingWindowLevel() + 1` hack.
-                let _: () = msg_send![
-                    *self.ns_window,
-                    setLevel: ffi::NSWindowLevel::NSNormalWindowLevel
-                ];
+                let _: () = msg_send![*self.ns_window, setLevel: ffi::kCGBaseWindowLevelKey];
             },
             _ => {}
         };
@@ -1101,14 +1089,12 @@ impl UnownedWindow {
 
     #[inline]
     pub fn focus_window(&self) {
-        let is_minimized: BOOL = unsafe { msg_send![*self.ns_window, isMiniaturized] };
-        let is_minimized = is_minimized == YES;
-        let is_visible: BOOL = unsafe { msg_send![*self.ns_window, isVisible] };
-        let is_visible = is_visible == YES;
+        let is_minimized: bool = unsafe { msg_send![*self.ns_window, isMiniaturized] };
+        let is_visible: bool = unsafe { msg_send![*self.ns_window, isVisible] };
 
         if !is_minimized && is_visible {
             unsafe {
-                NSApp().activateIgnoringOtherApps_(YES);
+                NSApp().activateIgnoringOtherApps_(Bool::YES.as_raw());
                 util::make_key_and_order_front_async(*self.ns_window);
             }
         }
@@ -1242,7 +1228,7 @@ impl WindowExtMacOS for UnownedWindow {
                 // Set the window frame to the screen frame size
                 let screen = self.ns_window.screen();
                 let screen_frame = NSScreen::frame(screen);
-                NSWindow::setFrame_display_(*self.ns_window, screen_frame, YES);
+                NSWindow::setFrame_display_(*self.ns_window, screen_frame, Bool::YES.as_raw());
 
                 // Fullscreen windows can't be resized, minimized, or moved
                 util::toggle_style_mask(
@@ -1257,7 +1243,7 @@ impl WindowExtMacOS for UnownedWindow {
                     NSWindowStyleMask::NSResizableWindowMask,
                     false,
                 );
-                NSWindow::setMovable_(*self.ns_window, NO);
+                NSWindow::setMovable_(*self.ns_window, Bool::NO.as_raw());
 
                 true
             } else {
@@ -1270,8 +1256,8 @@ impl WindowExtMacOS for UnownedWindow {
                 }
 
                 let frame = shared_state_lock.saved_standard_frame();
-                NSWindow::setFrame_display_(*self.ns_window, frame, YES);
-                NSWindow::setMovable_(*self.ns_window, YES);
+                NSWindow::setFrame_display_(*self.ns_window, frame, Bool::YES.as_raw());
+                NSWindow::setMovable_(*self.ns_window, Bool::YES.as_raw());
 
                 true
             }
@@ -1280,15 +1266,12 @@ impl WindowExtMacOS for UnownedWindow {
 
     #[inline]
     fn has_shadow(&self) -> bool {
-        unsafe { self.ns_window.hasShadow() == YES }
+        unsafe { Bool::from_raw(self.ns_window.hasShadow()).as_bool() }
     }
 
     #[inline]
     fn set_has_shadow(&self, has_shadow: bool) {
-        unsafe {
-            self.ns_window
-                .setHasShadow_(if has_shadow { YES } else { NO })
-        }
+        unsafe { self.ns_window.setHasShadow_(Bool::new(has_shadow).as_raw()) }
     }
 }
 
@@ -1316,14 +1299,14 @@ unsafe fn set_min_inner_size<V: NSWindow + Copy>(window: V, mut min_size: Logica
     // If necessary, resize the window to match constraint
     if current_rect.size.width < min_size.width {
         current_rect.size.width = min_size.width;
-        window.setFrame_display_(current_rect, NO)
+        window.setFrame_display_(current_rect, Bool::NO.as_raw())
     }
     if current_rect.size.height < min_size.height {
         // The origin point of a rectangle is at its bottom left in Cocoa.
         // To ensure the window's top-left point remains the same:
         current_rect.origin.y += current_rect.size.height - min_size.height;
         current_rect.size.height = min_size.height;
-        window.setFrame_display_(current_rect, NO)
+        window.setFrame_display_(current_rect, Bool::NO.as_raw())
     }
 }
 
@@ -1341,14 +1324,14 @@ unsafe fn set_max_inner_size<V: NSWindow + Copy>(window: V, mut max_size: Logica
     // If necessary, resize the window to match constraint
     if current_rect.size.width > max_size.width {
         current_rect.size.width = max_size.width;
-        window.setFrame_display_(current_rect, NO)
+        window.setFrame_display_(current_rect, Bool::NO.as_raw())
     }
     if current_rect.size.height > max_size.height {
         // The origin point of a rectangle is at its bottom left in Cocoa.
         // To ensure the window's top-left point remains the same:
         current_rect.origin.y += current_rect.size.height - max_size.height;
         current_rect.size.height = max_size.height;
-        window.setFrame_display_(current_rect, NO)
+        window.setFrame_display_(current_rect, Bool::NO.as_raw())
     }
 }
 
@@ -1360,8 +1343,8 @@ pub(super) fn get_ns_theme() -> Theme {
         ];
         let app_class = class!(NSApplication);
         let app: id = msg_send![app_class, sharedApplication];
-        let has_theme: BOOL = msg_send![app, respondsToSelector: sel!(effectiveAppearance)];
-        if has_theme == NO {
+        let has_theme: bool = msg_send![app, respondsToSelector: sel!(effectiveAppearance)];
+        if !Bool::from_raw(has_theme).as_bool() {
             return Theme::Light;
         }
         let appearance: id = msg_send![app, effectiveAppearance];
@@ -1389,8 +1372,8 @@ fn set_ns_theme(theme: Theme) {
     unsafe {
         let app_class = class!(NSApplication);
         let app: id = msg_send![app_class, sharedApplication];
-        let has_theme: BOOL = msg_send![app, respondsToSelector: sel!(effectiveAppearance)];
-        if has_theme == YES {
+        let has_theme: bool = msg_send![app, respondsToSelector: sel!(effectiveAppearance)];
+        if Bool::from_raw(has_theme).as_bool() {
             let name = NSString::alloc(nil).init_str(name);
             let appearance: id = msg_send![class!(NSAppearance), appearanceNamed: name];
             let _: () = msg_send![app, setAppearance: appearance];